LayoutTests/webgpu/render-passes.html

<!DOCTYPE html>
<html>
<script src="../resources/js-test-pre.js"></script>
<script src="js/basic-webgpu-functions.js"></script>
<script>
if (window.testRunner)
    window.testRunner.dumpAsText();

let commandBuffer, renderPassEncoder;

function setUpBasicRenderPassEncoder() {
    commandBuffer = defaultDevice.createCommandBuffer();

    // Default a loadOp, storeOp in the implementation for now.
    const colorAttachmentDescriptor = {
        attachment: context.getNextTexture().createDefaultTextureView(),
        clearColor: { r: 0.35, g: 0.65, b: 0.85, a: 1.0 }
    }

    let renderPassDescriptor = {
        colorAttachments: [colorAttachmentDescriptor]
    }

    renderPassEncoder = commandBuffer.beginRenderPass(renderPassDescriptor);
    if (renderPassEncoder)
        testPassed("Successfully created basic WebGPURenderPassEncoder.");
    else
        testFailed("Could not create WebGPURenderPassEncoder!");
}

function checkBadRenderPassDescriptor(descriptor, testSubjectName) {
    renderPassEncoder = commandBuffer.beginRenderPass(descriptor);

    if (renderPassEncoder)
        testFailed(`WebGPURenderPassEncoder was created with an invalid ${testSubjectName}!`);
    else 
        testPassed(`WebGPURenderPassEncoder with invalid ${testSubjectName} was not created.`);
}

function noAttachmentDescriptors() {
    checkBadRenderPassDescriptor(null, "WebGPURenderPassDescriptor");
    checkBadRenderPassDescriptor({}, "WebGPURenderPassDescriptor");
}

runWebGPUTests([setUpBasicRenderPassEncoder, noAttachmentDescriptors]);

successfullyParsed = true;
</script>
<script src="../resources/js-test-post.js"></script>
</html>
